requirements.txt
GitPython==2.1.11
requests==2.9.1
